Server cleanup

+++ code/client/src/views/user/home.vue
@@ -9,7 +9,7 @@
                 :backgroundColor="`none`"
                 :basedOn="{ username: $store.getters.decodedJWT.username }"
                 :buttonText="`Change Username`"
-                @submitted="changeUsername($event)"
+                @submitted="updateInfo($event)"
             />
         </collapsable-box>
         <collapsable-box
@@ -21,7 +21,7 @@
                 :bindedExts="['confirms']"
                 :backgroundColor="`none`"
                 :buttonText="`Change Password`"
-                @submitted="changePassword($event)"
+                @submitted="updateInfo($event)"
             />
         </collapsable-box>
         <collapsable-box
@@ -33,7 +33,7 @@
                 :basedOn="$store.getters.decodedJWT"
                 :backgroundColor="`none`"
                 :buttonText="`Update Profile`"
-                @submitted="updateProfile($event)"
+                @submitted="updateInfo($event)"
             />
         </collapsable-box>
         <collapsable-box
@@ -46,7 +46,7 @@
                 v-if="khateebs.struct"
                 :options="khateebs.struct"
                 :currentlySelected="khateebs.availableTimings"
-                @changed="modifyAvailableTimings($event)"
+                @changed="updateInfo({ availableTimings: $event }, false)"
             />
         </collapsable-box>
         <collapsable-box
@@ -56,7 +56,7 @@
         >
             <calendar
                 :originalVal="$store.getters.decodedJWT.unavailableDates" 
-                @changed="modifyUnavailableDates($event)"
+                @changed="updateInfo($event, false)"
             />
         </collapsable-box>
         <collapsable-box
@@ -135,37 +135,14 @@ export default {
         }
     },
     methods: {
-        async changePassword($event) {
+        async updateInfo($event, withMsgAndPush=true) {
             try {
-                const res = await this.$API.user.changePassword($event)
-                this._.alert(`Successfully changed password`, 'success')
-                this._.toHomePage()
-            } catch(err) {
-                console.log(err)
-            }
-        },
-        async changeUsername($event) {
-            try {
-                const res = await this.$API.user.changeUsername($event)
+                const res = await this.$API.user.updateInfo($event)
                 this.storeToken(res.token)
-                this._.alert(`Successfully changed username`, 'success')
-                this._.toHomePage()
-            } catch(err) {
-                console.log(err)
-            }
-        },
-        async updateProfile($event) {
-            try {
-                const updated = {
-                    firstName: $event.firstName,
-                    lastName: $event.lastName,
-                    phoneNumber: $event.phoneNumber,
-                    handle: $event.handle
+                if (withMsgAndPush) {
+                    this._.alert(`Successfully updated!`, 'success')
+                    this._.toHomePage()
                 }
-                const res = await this.$API.user.changeProfile(updated)
-                this.storeToken(res.token)
-                this._.alert(`Successfully updated profile`, 'success')
-                this._.toHomePage()
             } catch(err) {
                 console.log(err)
             }
@@ -178,11 +155,13 @@ export default {
         async deleteAccount() {
             try {
                 const confirm = await this._.confirm(`Are you sure you want to permenantly delete your account?`)
+                if (!confirm)
+                    return
                 const res = await this.$API.user.deleteAccount()
+                console.log(res)
                 this.$store.dispatch('logout')
                 this._.toHomePage()
                 this._.alert(`Successfully delete account`, 'success')
-                console.log(res)
             } catch(err) {
                 console.log(err)
             }
@@ -223,14 +202,6 @@ export default {
                 console.log(err)
             }
         },
-        async modifyUnavailableDates($event) {
-            try {
-                const res = await this.$API.user.changeProfile({ unavailableDates: $event })
-                this.storeToken(res.token)
-            } catch(err) {
-                console.log(err)
-            }
-        }
     },
     computed: {
         availableTimingsTag() {
